CODE-16: Reduce calls to recalculate anchor date by explicitly giving date in created menu
Note that with validation enforced that timezone becomes a requirement.
UTC is used because of the Todoist API, where the string is interpreted
as UTC time, despite a user having a default regional time. If it were
beneficial, we could at a later step convert from European to UTC, etc,
but for now, stick with UTC.

https://developer.todoist.com/rest/v2/#create-a-new-task