CODE-214 Add basic ingredient selection
This filter is not very robust. However, we really need
to handle the data quality issues first. For instance,
it could be that we are looking for 'penne' which
isn't a true ingredient and is replaced with 'fusili'.
We'll get a warning that it can't be found or maybe in
some improper recipes it is found. Additionally, there
are some ingredients whose names are not unique (red
cabbage vs cabbage); if we search fo cabbage meaning
green cabbage, we might end up with some red cabbage
options.